Modify internal packages structure.
