+++ src/main/java/org/nkjmlab/sorm4j/extension/logger/package.html
@@ -1,0 +1,5 @@
+<html>
+<body>
+  <p>Provides Logger.</p>
+</body>
+</html>
