+++ src/main/java/org/nkjmlab/sorm4j/internal/util/DriverManagerDataSource.java
@@ -1,4 +1,4 @@
-package org.nkjmlab.sorm4j.internal.mapping;
+package org.nkjmlab.sorm4j.internal.util;
 
 import java.io.PrintWriter;
 import java.sql.Connection;
@@ -7,7 +7,6 @@ import java.sql.SQLException;
 import java.sql.SQLFeatureNotSupportedException;
 import java.util.logging.Logger;
 import javax.sql.DataSource;
-import org.nkjmlab.sorm4j.internal.util.Try;
 
 /**
  * A database connection source wrapped a {@link DriverManager}
