+++ src/test/java/org/nkjmlab/sorm4j/internal/mapping/TableMappingTest.java
@@ -12,6 +12,7 @@ import org.nkjmlab.sorm4j.Sorm;
 import org.nkjmlab.sorm4j.common.Guest;
 import org.nkjmlab.sorm4j.common.Player;
 import org.nkjmlab.sorm4j.common.SormTestUtils;
+import org.nkjmlab.sorm4j.internal.OrmConnectionImpl;
 
 class TableMappingTest {
   private Sorm sorm;
