+++ src/test/java/org/nkjmlab/sorm4j/internal/mapping/multirow/OrmConfigStoreTest.java
@@ -10,9 +10,9 @@ import org.nkjmlab.sorm4j.extension.DefaultTableNameMapper;
 import org.nkjmlab.sorm4j.extension.MultiRowProcessorType;
 import org.nkjmlab.sorm4j.extension.ResultSetConverter;
 import org.nkjmlab.sorm4j.extension.SormConfig;
-import org.nkjmlab.sorm4j.extension.SormContext;
 import org.nkjmlab.sorm4j.extension.SqlParametersSetter;
 import org.nkjmlab.sorm4j.extension.TableNameMapper;
+import org.nkjmlab.sorm4j.internal.SormContext;
 
 class OrmConfigStoreTest {
